+++ R/eem_utils.R
@@ -1,26 +1,17 @@
 #' @importFrom grDevices colorRampPalette
 #' @importFrom graphics par plot text
+#' @importFrom viridis viridis
 .plot_eem <- function(x, show_peaks, ...){
 
 
-  jet.colors <- colorRampPalette(c("#00007F",
-                                   "blue",
-                                   "#007FFF",
-                                   "cyan",
-                                   "#7FFF7F",
-                                   "yellow",
-                                   "#FF7F00",
-                                   "red",
-                                   "#7F0000"))
-
-  fields::image.plot(y = x$em,
+fields::image.plot(y = x$em,
              x = x$ex,
              z = t(x$x),
              main = paste(x$sample, "\n", attr(x, "manucafturer"), sep = ""),
              xlab = "Excitation (nm.)",
              ylab = "Emission (nm.)",
              legend.lab = "Fluorescence intensity",
-             col = jet.colors(255),
+             col = viridis::viridis(256),
              ...)
 
   if(show_peaks){